Implemented custom chat window UI
