+++ BotsSDK/korebot/src/main/java/kore/botssdk/activity/BotChatActivity.java
@@ -69,8 +69,8 @@ import kore.botssdk.event.KoreEventCenter;
 import kore.botssdk.events.SocketDataTransferModel;
 import kore.botssdk.fileupload.core.KoreWorker;
 import kore.botssdk.fileupload.core.UploadBulkFile;
-import kore.botssdk.fragment.BotContentFragment;
-import kore.botssdk.fragment.ComposeFooterFragment;
+import kore.botssdk.fragment.content.BotContentFragment;
+import kore.botssdk.fragment.footer.ComposeFooterFragment;
 import kore.botssdk.listener.BaseSocketConnectionManager;
 import kore.botssdk.listener.BotContentFragmentUpdate;
 import kore.botssdk.listener.BotSocketConnectionManager;
