+++ BotsSDK/korebot/src/main/java/kore/botssdk/net/SDKConfig.java
@@ -4,7 +4,13 @@ import android.view.View;
 
 import androidx.annotation.NonNull;
 
+import kore.botssdk.fragment.content.BaseContentFragment;
+import kore.botssdk.fragment.footer.BaseFooterFragment;
+
 public class SDKConfig {
+    private static BaseContentFragment customContentFragment = null;
+    private static BaseFooterFragment customFooterFragment = null;
+
     public static void setCustomTemplateView(@NonNull String templateName, @NonNull View templateView) {
         SDKConfiguration.setCustomTemplateView(templateName, templateView);
     }
@@ -37,5 +43,17 @@ public class SDKConfig {
     public static void setJwtTokenUrl(String url) {
         SDKConfiguration.JWTServer.setJwtServerUrl(url);
     }
+
+    public static void  addCustomContentFragment(BaseContentFragment contentFragment) {
+        customContentFragment = contentFragment;
+    }
+
+    public static BaseContentFragment getCustomContentFragment()  {return customContentFragment;}
+
+    public static void addCustomFooterFragment(BaseFooterFragment fragment) {
+        customFooterFragment = fragment;
+    }
+
+    public static BaseFooterFragment getCustomFooterFragment() {return customFooterFragment;}
 
 }
